Data.Chain

Synopsis

Chain node edge is a linear chain of nodes with directed edges.

Chain

data Chain node edge

A linear chain of nodes. Edges between nodes are labeled by a Monoid edge.

  n_tip  <--e_tip-- … <--e1-- n1 <--e0-- n0

member :: Ord node => node -> Chain node edge -> Bool

Test whether a node is contained in the chain.

type ChainContext node edge = Edge (Maybe (edge, node)) node

Context (incoming and outgoing edges) for a node in a Chain.

lookup :: Ord node => node -> Chain node edge -> Maybe (ChainContext node edge)

Look up the Context of a node in a Chain.

fromEdge :: Ord node => Edge node edge -> Chain node edge

Construct a chain from a single Edge.

fromEdges :: Ord node => [Edge node edge] -> Maybe (Chain node [edge])

Construct a chain from a collection of edges. Fails if the edges do not fit together.

The ordering of edge labels of a single edge in the chain will be the same as the ordering of the edge labels as they appear in the list.

edges :: Ord node => Chain node edge -> [edge]

List all edges in the Chain.

The edge that points to the tip is listed first, and the edge that starts at the beginning is listed last.

nodes :: Ord node => Chain node edge -> [node]

List all nodes in the Chain. The tip is listed first.

toEdges :: Chain node edge -> [Edge node edge]

Convert a Chain into a list of Edge.

summary :: (Ord node, Monoid edge) => Chain node edge -> edge

Combine all the edges in the Chain. The summary is invariant under collapseNode.

summary = mconcat . edges

DeltaChain

data DeltaChain node edge

Changes to a Chain.

Constructors

AppendTip node edge
See appendTip.

CollapseNode node
See collapseNode.

RollbackTo node
See rollbackTo.

appendTip :: Ord node => node -> edge -> Chain node edge -> Chain node edge

Append a new tip to the chain.

collapseNode :: (Ord node, Semigroup edge) => node -> Chain node edge -> Chain node edge

Remove the given node and combine the incoming and outgoing edges. Do nothing if the node is at the tip, or at the bottom, or not in the chain at all.

rollbackTo :: Ord node => node -> Chain node edge -> Chain node edge

Remove the tip and more nodes from the chain until the given node is the tip.

Do nothing if the node is not in the chain.

chainIntoTable :: (Ord node, Semigroup edge) => (edge -> Pile e) -> (Pile e -> edge) -> Embedding (DeltaChain node edge) [DeltaTable (Edge node e)]

Embed a Chain into a table of Edge.

The first and second argument specify how the edge labels are to be mapped to and from sets of table rows. Importantly, we may not assume that the table stores the rows in any particular order.

data ErrMalformedChainTable

Constructors

ErrMalformedChainTable

Edge

data Edge node edge

Utility type that represents an Edge in a graph: it connects two node via an edge label.

Constructors

Edge

Fields

from :: node

to :: node

via :: edge

flattenEdge :: Edge node [edge] -> [Edge node edge]

Flatten a list of edges

Testing

testChain :: (Table (Edge Int Char), [[DeltaDB Int (Edge Int Char)]])