Data.Delta

Synopsis

Delta encodings.

The type constraint Delta delta means that the type delta is a delta encoding of the corresponding base type Base delta.

Delta encodings can be transformed into each other using an Embedding.

Delta encodings

class Delta delta where

Type class for delta encodings.

Associated Types

type Base delta :: Type

Base type for which delta represents a delta encoding. This is implemented as a type family, so that we can have multiple delta encodings for the same base type.

FIXME: Better name for Base? It's sooo generic. Pier, dock, ground, anchor, site, …

Methods

apply :: delta -> Base delta -> Base delta

Apply a delta encoding to the base type.

data NoChange a

Trivial delta encoding for the type a that admits no change at all.

Constructors

NoChange

newtype Replace a

Trivial delta encoding for the type a that replaces the value wholesale.

Constructors

Replace a

data DeltaList a

Delta encoding for lists where a list of elements is prepended.

Constructors

Append [a]

data DeltaSet1 a

Delta encoding for Set where a single element is deleted or added.

Constructors

Insert a

Delete a

data DeltaSet a

Delta encoding for a Set where collections of elements are inserted or deleted.

mkDeltaSet :: Ord a => Set a -> Set a -> DeltaSet a

Delta to get from the second argument to the first argument.

deltaSetToList :: DeltaSet a -> [DeltaSet1 a]

Flatten a DeltaSet to a list of DeltaSet1.

In the result list, the set of a appearing as Insert a is disjoint from the set of a appearing as Delete a.

deltaSetFromList :: Ord a => [DeltaSet1 a] -> DeltaSet a

Collect insertions or deletions of elements into a DeltaSet.

To save space, combinations of Insert and Delete for the same element are simplified when possible. These simplifications always preserve the property

apply (deltaSetFromList ds) = apply ds

Embedding of types and delta encodings

An Embedding da db embeds one type and its delta encoding da into another type and its delta encoding db.

For reasons of efficiency, Embedding is an abstract type. It is constructed using the Embedding' type, which has three components.

write embeds values from the type a = Base da into the type b = Base bd.

load attempts to retrieve the value of type a from the type b, but does not necessarily succeed.

update maps a delta encoding da to a delta encoding db. For this mapping, the value of type a and a corresponding value of type b are provided; the delta encodings of types da and db are relative to these values.

The embedding of one type into the other is characterized by the following properties:

The embedding need not be surjective: The type b may contain many values that do not correspond to a value of type a. Hence, load has an Either result. (See Note [EitherSomeException] for the choice of exception type.) However, retrieving a written value always succeeds, we have

load . write = Right

The embedding is redundant: The type b may contain multiple values that correspond to one and the same a. This is why the update function expects the type b as argument, so that the right delta encoding can be computed. Put differently, we often have

write a ≠ b   where Right a = load b

The embedding of delta encoding commutes with apply. We have

Just (apply da a) = load (apply (update a b da) b)
    where Right a = load b

However, since the embedding is redundant, we often have

apply (update a (write a) da) (write a) ≠ write (apply da a)

data Embedding da db

Embedding with efficient composition o. To construct an embedding, use mkEmbedding.

class Semigroupoid (c :: k -> k -> Type) where

Category sans id

Methods

o :: forall (j :: k) (k1 :: k) (i :: k). c j k1 -> c i j -> c i k1

data Embedding' da db where

Specification of an embedding of a type a with delta encoding da into the type b with delta encoding db. See the discussion of Embedding for a more detailed description.

Constructors

Embedding'

Fields

:: (Delta da, Delta db, a ~ Base da, b ~ Base db)

=> { load :: b -> Either SomeException a

, write :: a -> b

, update :: a -> b -> da -> db

} -> Embedding' da db

mkEmbedding :: Embedding' da db -> Embedding da db

Construct Embedding with efficient composition

fromEmbedding :: (Delta da, Delta db) => Embedding da db -> Embedding' da db

Extract load, write, and update functions from an efficient Embedding.

pair :: Embedding da1 db1 -> Embedding da2 db2 -> Embedding (da1, da2) (db1, db2)

A pair of Embeddings gives an embedding of pairs.

liftUpdates

Arguments

:: Delta da

=> Embedding da db

-> [da]
List of deltas to apply. The head is applied last.

-> Base da
Base value to apply the deltas to.

-> (Base db, [db])
(Final base value, updates that were applied (head is last)).

Lift a sequence of updates through an Embedding.

replaceFromApply :: (Delta da, a ~ Base da) => Embedding' da (Replace a)

Having an apply function is equivalent to the existence of a canonical embedding into the trivial Replace delta encoding.

Internal

inject :: Embedding da db -> Base da -> Machine da db

project :: Embedding da db -> Base db -> Either SomeException (Base da, Machine da db)

data Machine da db

Strict pair. If a value of this type is in WHNF, so are the two components. data StrictPair a b = !a :*: !b infixr 1 :*:

A state machine that maps deltas to deltas. This machine always carries a state of type Base db around.

Constructors

Machine

Fields

state_ :: !(Base db)

step_ :: (Base da, da) -> (db, Machine da db)

idle :: Delta da => Base da -> Machine da da

Identity machine starting from a base type.

pairMachine :: Machine da1 db1 -> Machine da2 db2 -> Machine (da1, da2) (db1, db2)

Pair two Machine.

fromState :: Delta db => ((Base da, da) -> (Base db, s) -> (db, s)) -> (Base db, s) -> Machine da db

Create a Machine from a specific state s, and the built-in state Base db.