Data.Table

Synopsis

Table models a database table. It corresponds to a collection of rows. Each row has a unique ID, but this is transparent to the API user.

Pile models a set of values. Unlike Set, it is represented as a lightweight list. This is used to highlight that the ordering of rows in a Table is not deterministic.

Supply is a supply of unique IDs.

Table

data Table row

A Table is a collection of rows.

Constructors

Table

Fields

rows :: IntMap row

Rows indexed by unique ID.

uids :: Supply

Unique ID supply. WARNING: This is an internal part of the structure. Changing it may lead to an inconsistent state.

empty :: Table row

The empty Table, containing no rows.

fromRows :: [(Int, row)] -> Table row

Construct a Table from a list of rows with unique IDs.

fromList :: [row] -> Table row

Construct a Table from a list of rows

toPile :: Table row -> Pile row

Pile of rows contained in the Table.

toRows :: Table row -> Pile (Int, row)

Pile of rows with unique IDs contained in the Table.

selectWhere :: (row -> Bool) -> Table row -> Pile row

List all rows satisfying the predicate.

insertMany :: [row] -> Table row -> Table row

Insert rows into the table.

deleteWhere :: (row -> Bool) -> Table row -> Table row

Delete all rows satisfying the predicate.

updateWhere :: (row -> Bool) -> (row -> row) -> Table row -> Table row

Update all rows satisfying the predicate

data DeltaTable row

Delta encoding for changes to a Table.

Constructors

InsertMany [row]

DeleteWhere (row -> Bool)

UpdateWhere (row -> Bool) (row -> row)

data DeltaDB key row

Delta encoding for changes to a database table with unique IDs.

Constructors

InsertManyDB [(key, row)]

DeleteManyDB [key]

UpdateManyDB [(key, row)]

tableIntoDatabase :: Embedding [DeltaTable row] [DeltaDB Int row]

Pile

newtype Pile a

A Pile is a set of values. Unlike Set, it is represented as a list, and avoids the Ord constraint.

This type is useful for highlighting that a collection of values has no specific order, even though it is not represented as a Set.

Constructors

Pile

Fields

getPile :: [a]

fromSet :: Set a -> Pile a

deltaListToPile :: DeltaList a -> Pile (Int, a)

Map a DeltaList to a Pile of indexed single element concatenations. Higher indices are prepended later.

deltaListFromPile :: Pile (Int, a) -> DeltaList a

Restore a DeltaList from a Pile.

deltaListFromPile . deltaListToPile = id

deltaSetToPile :: DeltaSet a -> Pile (DeltaSet1 a)

Randomly permute the objects in a Pile. Useful for stress testing.

Every function f :: Pile A -> B should satisfy

forall g.  f . permute g = f

permute :: RandomGen g => g -> Pile a -> Pile a permute = undefined let (index, g2) = randomR (1,n) g1

Map a DeltaSet to a Pile of single element insertions and deltions.

deltaSetFromPile :: Ord a => Pile (DeltaSet1 a) -> DeltaSet a

Restore a DeltaSet from a Pile of single element insertions and deletions.

deltaSetFromPile . deltaSetToPile = id

Supply

data Supply

A supply of unique IDs.

abundance :: Supply

Fresh supply of unique IDs.

fresh :: Supply -> (Int, Supply)

Retrieve a fresh unique ID.

consume :: [Int] -> Supply -> Supply

Remove a list of unique IDs from the Supply if necessary.