Database.Persist.Delta

Synopsis

Manipulating SQL database tables using delta encodings via the "persistent" package.

Store

newEntityStore :: forall row m. (PersistRecordBackend row SqlBackend, ToBackendKey SqlBackend row, Show row, MonadIO m) => m (Store SqlPersistM [DeltaDB Int row])

Construct a Store for Entity.

FIXME: This function should also do "migrations", i.e. create the database table in the first place.

newSqlStore :: (MonadIO m, IsRow row, IsRow (row :. Col "id" Primary), Show row) => m (Store SqlPersistM [DeltaDB Int row])

Construct a Store from an SQL table.

The unique IDs will be stored in a column "id" at the end of each row in the database table.