Database.Schema

Synopsis

Typed database tables and rows.

Database columns

class PersistField a => IsCol a where

Class of columns that can be stored in database tables.

Methods

getSqlType :: Proxy a -> SqlColType

newtype SqlColType

SQL column types, including constraints. Values of type SqlColType represent SQL types such as

INTEGER  PRIMARY KEY NOT NULL
TEXT  NOT NULL

Constructors

SqlColType Text

toColType :: SqlType -> SqlColType

Helper for converting SqlType into an SQL column type with constraints.

newtype Primary

Primary key.

Constructors

Primary

Fields

getPrimary :: Int

Database rows and tables

data Table (name :: Symbol)

Named database table.

Constructors

Table

newtype Col (name :: Symbol) a

Named database column.

Constructors

Col a

data a :. b infixl 3

Infix notation for a pair of types.

Constructors

a :. b infixl 3

class IsRow row where

Class of row types that can be stored in database tables. Instances of this class are essentially lists of columns. Example:

type PersonRow = Table "person" :. Col "name" Text :. Col "age" Int

Methods

getTableName :: Proxy row -> Text

getColNames :: Proxy row -> [Text]

getSqlTypes :: Proxy row -> [SqlColType]

toSqlValues :: row -> [PersistValue]

fromSqlValues :: [PersistValue] -> Either Text row

SQL Queries

data Query row

An SQL query that returns a list of values of type row.

callSql :: (MonadIO m, IsRow row) => Query row -> SqlPersistT m [row]

Run an SQL query and return a list of rows as result.

runSql :: MonadIO m => Query () -> SqlPersistT m ()

Execute an SQL query, but do not return any results

createTable :: IsRow row => Proxy row -> Query ()

Create a database table that can store the given rows.

selectAll :: forall row. IsRow row => Query row

Select all rows from the table.

insertOne :: forall row. IsRow row => row -> Query ()

Insert a single row into the corresponding table.

repsertOne :: forall row. IsRow row => (row :. Col "id" Primary) -> Query ()

Replace or insert a single row with a primary key into a database.

FIXME: It would be nicer if the "id" column was the first column instead of the last column in the table.

updateOne :: forall row. IsRow row => (row :. Col "id" Primary) -> Query ()

Update one row with a given "id" column in a database table.

deleteAll :: forall row. IsRow row => Proxy row -> Query ()

Delete all rows in a database table

deleteOne :: forall row. IsRow row => Proxy row -> Col "id" Primary -> Query ()

Delete one row with a given "id" column in a database table.

Testing

testPerson :: PersonRow